feat: Extract the logic of DataLoaderView into a controller to ease reuse.
